rune/rune/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.7
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
filelock==3.13.1
importlib-metadata==6.8.0
kombu==5.3.4
Markdown==3.5.1
pipenv==2023.11.15
platformdirs==4.0.0
prompt-toolkit==3.0.41
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
virtualenv==20.24.7
wcwidth==0.2.12
zipp==3.17.0
